+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/).
 - Invalid configuration submit bug. #1301
 - Notification spam bug. #2731
 - Agent propagator crashes if exploiters malfunction. #2992
+- Configuration order not preserved in debugging output. #2860
 
 ### Security
 - Fixed plaintext private key in SSHKey pair list in UI. #2950